+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fixed hook listener discovery
+
 ## v0.2.1 (2024-08-01)
 * Fixed Hook index loader